requirements.txt
Django==5.2.4
gunicorn==23.0.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.9.0
django-environ==0.11.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
asgiref==3.9.1

